+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add a cert-manager ClusterIssuer readiness test.
+
 ## [1.11.1] - 2023-08-31
 
 ### Changed